refactor into base class
